LayoutTests/discovery/RUIHPL-7.9.3.1.4.b.html

<html>

<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>

<body>
<script>
var vid;
var fileName = '/home/gar/MediaContent/SUPER_test5_LPCM_Mono_VBR_16SS_44100Hz.WAV';
var canplay = false;

function playing()
{
	testPassed("Audio Track is playing.");
}

function paused()
{
	testPassed("Audio Track is paused.");
}

function playable() {
	canplay = true;
	vid = document.getElementsByTagName("audio")[0];
	vid.audioTracks.oncanplay = null;
	testPassed("Audio Track is playable: " + fileName);
	
	finishJSTest();
	
}

function notPlayable()
{
	if (caplay)
		return;
	
	testFailed("Audio Track is NOT playable: " + fileName);
}

function insertCtrl() {
	var body = document.getElementsByTagName("body")[0];
	var v = document.createElement("audio");
	v.setAttribute('controls', 'yes');
	v.setAttribute('width', '740px');
	v.addEventListener("canplay", playable, true);
	var src = document.createElement("source");
	src.setAttribute('src', fileName);
	v.appendChild(src);
	body.appendChild(v);
	setTimeout(notPlayable,4000);
}

insertCtrl();
description('Tests DLNA Player for AV Media Format. RUIHPL-7.9.3.1.2: LPCM');
window.jsTestIsAsync = true;
</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>

</html>
