LayoutTests/discovery/RUIHPL-7.9.3.1.4.b-expected.txt

Tests DLNA Player for AV Media Format. RUIHPL-7.9.3.1.2: LPCM

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Audio Track is playable: /home/gar/MediaContent/SUPER_test5_LPCM_Mono_VBR_16SS_44100Hz.WAV
PASS successfullyParsed is true

TEST COMPLETE
